Rename tokenizer module
